python/requirements.txt
opencv-python
numpy
torch
torchvision
tqdm
pillow
matplotlib